调整了一下模型输出内容，换了一下openvino和tensorrt的转化器位置

+++ nn/src/main/python/infer/qwen/source_infer.py
@@ -36,7 +36,7 @@ class QwenMoelRun(BaseMoelRun):
         with torch.no_grad():
             output = self.model(torch.from_numpy(inputs["input_ids"]).to(self.device),torch.from_numpy(inputs["attention_mask"]).to(self.device),torch.from_numpy(inputs["position_ids"]).to(self.device),past_key_values)
  
-            return output[0].cpu().detach().numpy(),output[1]
+            return output[0].cpu().detach().to(torch.float32).numpy(),output[1]
 
 if __name__ == "__main__":
     qwen = QwenMoelRun()
